+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Enhancements:
 * venv to .gitignore
 * move terraform.tfvars.json to root of repo
 * move libs/py/utils/test_logger.py to libs/py/utils/tests.py to have the same test rules approach for all libs
+* `terraform/tf_variables.tf.tpl` reconfigure defaults for `env.cloud.buckets` and `env.kubernetes`
 
 # v0.6.0